requirements.txt
cycler==0.11.0
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
neurokit2==0.1.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
